Try scan config
