+++ app/sampledata/dog/12228_Dog_v1_L2.sfa
@@ -1,0 +1,53 @@
+{
+  materials: [
+    {
+      name: 'Dog',
+      parameters: [
+        {
+          baseColor: 'Dog_diffuse',
+        },
+        {
+          baseColorTint: [
+            1,
+            1,
+            1,
+            1,
+          ],
+        },
+        {
+          metallic: 1,
+        },
+        {
+          roughness: 0.408248,
+        },
+        {
+          opacity: null,
+        },
+      ],
+      source: 'build/sceneform_sdk/default_materials/obj_material.sfm',
+    },
+  ],
+  model: {
+    attributes: [
+      'Position',
+      'TexCoord',
+      'Orientation',
+    ],
+    collision: {},
+    file: 'sampledata/dog/12228_Dog_v1_L2.obj',
+    name: '12228_Dog_v1_L2',
+    recenter: 'root',
+    scale: 0.074973999999999999,
+  },
+  samplers: [
+    {
+      file: 'sampledata/dog/Dog_diffuse.jpg',
+      name: 'Dog_diffuse',
+      params: {
+        usage_type: 'Data',
+      },
+      pipeline_name: 'Dog_diffuse.jpg',
+    },
+  ],
+  version: '0.54:2',
+}
