+++ app/sampledata/models/frame_lower_left.sfa
@@ -1,0 +1,50 @@
+{
+  materials: [
+    {
+      name: 'unlit_material',
+      parameters: [
+        {
+          baseColor: 'frame_base',
+        },
+        {
+          baseColorTint: [
+            1,
+            1,
+            1,
+            1,
+          ],
+        },
+        {
+          metallic: 1,
+        },
+        {
+          roughness: 1,
+        },
+        {
+          opacity: null,
+        },
+      ],
+      source: 'build/sceneform_sdk/default_materials/obj_material.sfm',
+    },
+  ],
+  model: {
+    attributes: [
+      'Position',
+      'TexCoord',
+      'Orientation',
+    ],
+    collision: {},
+    file: 'sampledata/models/frame_lower_left.obj',
+    name: 'frame_lower_left',
+    recenter: 'root',
+    scale: 1.0006200000000001,
+  },
+  samplers: [
+    {
+      file: 'sampledata/models/frame_base.png',
+      name: 'frame_base',
+      pipeline_name: 'frame_base.png',
+    },
+  ],
+  version: '0.54:2',
+}
