+++ app/src/main/java/com/lt3d/MainActivity.java
@@ -42,13 +42,6 @@ public class MainActivity extends AppCompatActivity {
     BottomNavigationView navView;
     private static final int RC_SIGN_IN = 123;
 
-    private ArFragment arFragment;
-    private ImageView fitToScanView;
-
-    // Augmented image and its associated center pose anchor, keyed by the augmented image in
-    // the database.
-    private final Map<AugmentedImage, AugmentedImageNode> augmentedImageMap = new HashMap<>();
-
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -116,8 +109,7 @@ public class MainActivity extends AppCompatActivity {
                     changeFragment(new LibraryFragment());
                     return true;
                 case R.id.navigation_scan:
-//                    changeFragment(new ScanFragment());
-                    arFragmentConfig();
+                    changeFragment(new ScanFragment());
                     return true;
                 case R.id.navigation_setting:
                     changeFragment(new SettingFragment());
@@ -133,58 +125,10 @@ public class MainActivity extends AppCompatActivity {
         transaction.commit();
     }
 
-    private void arFragmentConfig() {
-        LayoutInflater factory = LayoutInflater.from(this);
-        View view = factory.inflate(R.layout.fragment_scan, null);
-
-        arFragment = (ArFragment) getSupportFragmentManager().findFragmentById(R.id.ux_fragment);
-        fitToScanView = view.findViewById(R.id.image_view_fit_to_scan);
-        arFragment.getArSceneView().getScene().addOnUpdateListener(this::onUpdateFrame);
-    }
-
     @Override
     protected void onPause() {
         super.onPause();
         overridePendingTransition(0, 0);
-    }
-
-    private void onUpdateFrame(FrameTime frameTime) {
-        Frame frame = arFragment.getArSceneView().getArFrame();
-
-        // If there is no frame or ARCore is not tracking yet, just return.
-        if (frame == null || frame.getCamera().getTrackingState() != TrackingState.TRACKING) {
-            return;
-        }
-
-        Collection<AugmentedImage> updatedAugmentedImages =
-                frame.getUpdatedTrackables(AugmentedImage.class);
-        for (AugmentedImage augmentedImage : updatedAugmentedImages) {
-            switch (augmentedImage.getTrackingState()) {
-                case PAUSED:
-                    // When an image is in PAUSED state, but the camera is not PAUSED, it has been detected,
-                    // but not yet tracked.
-                    String text = "Detected Image " + augmentedImage.getIndex();
-                    Alert.show(this, text);
-                    break;
-
-                case TRACKING:
-                    // Have to switch to UI Thread to update View.
-                    fitToScanView.setVisibility(View.GONE);
-
-                    // Create a new anchor for newly found images.
-                    if (!augmentedImageMap.containsKey(augmentedImage)) {
-                        AugmentedImageNode node = new AugmentedImageNode(this);
-                        node.setImage(augmentedImage);
-                        augmentedImageMap.put(augmentedImage, node);
-                        arFragment.getArSceneView().getScene().addChild(node);
-                    }
-                    break;
-
-                case STOPPED:
-                    augmentedImageMap.remove(augmentedImage);
-                    break;
-            }
-        }
     }
 
     public FirebaseUser getCurrentUser() {
