+++ app/src/main/java/com/lt3d/fragment/ScanFragment.java
@@ -1,143 +1,90 @@
 package com.lt3d.fragment;
 
-import android.app.ActivityManager;
-import android.content.Context;
-import android.content.res.AssetManager;
-import android.graphics.Bitmap;
-import android.graphics.BitmapFactory;
-import android.os.Build;
 import android.os.Bundle;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.ImageView;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 
-import com.firebase.ui.auth.AuthUI;
-import com.google.ar.core.AugmentedImageDatabase;
-import com.google.ar.core.Config;
-import com.google.ar.core.Session;
+import com.google.ar.core.AugmentedImage;
+import com.google.ar.core.Frame;
+import com.google.ar.core.TrackingState;
+import com.google.ar.sceneform.FrameTime;
 import com.google.ar.sceneform.ux.ArFragment;
 import com.lt3d.R;
 import com.lt3d.tools.Alert;
+import com.lt3d.tools.AugmentedImageNode;
 
-import java.io.IOException;
-import java.io.InputStream;
-import java.util.Arrays;
-import java.util.List;
+import java.util.Collection;
+import java.util.HashMap;
+import java.util.Map;
 
-public class ScanFragment extends ArFragment {
-    private static final String TAG = "AugmentedImageFragment";
+public class ScanFragment extends Fragment {
+    private ArFragment arFragment;
+    private ImageView fitToScanView;
+    private View view;
 
-    // This is the name of the image in the sample database.  A copy of the image is in the assets
-    // directory.  Opening this image on your computer is a good quick way to test the augmented image
-    // matching.
-    private static final String DEFAULT_IMAGE_NAME = "default.jpg";
+    // Augmented image and its associated center pose anchor, keyed by the augmented image in
+    // the database.
+    private final Map<AugmentedImage, AugmentedImageNode> augmentedImageMap = new HashMap<>();
 
-    // This is a pre-created database containing the sample image.
-    private static final String SAMPLE_IMAGE_DATABASE = "sample_database.imgdb";
-
-    // Augmented image configuration and rendering.
-    // Load a single image (true) or a pre-generated image database (false).
-    private static final boolean USE_SINGLE_IMAGE = false;
-
-    // Do a runtime check for the OpenGL level available at runtime to avoid Sceneform crashing the
-    // application.
-    private static final double MIN_OPENGL_VERSION = 3.0;
-
+    @Nullable
     @Override
-    public void onAttach(Context context) {
-        super.onAttach(context);
-
-        // Check for Sceneform being supported on this device.  This check will be integrated into
-        // Sceneform eventually.
-        if (Build.VERSION.SDK_INT < Build.VERSION_CODES.N) {
-            Log.e(TAG, "Sceneform requires Android N or later");
-            Alert.show(getContext(), "Sceneform requires Android N or later");
-        }
-
-        String openGlVersionString =
-                ((ActivityManager) context.getSystemService(Context.ACTIVITY_SERVICE))
-                        .getDeviceConfigurationInfo()
-                        .getGlEsVersion();
-        if (Double.parseDouble(openGlVersionString) < MIN_OPENGL_VERSION) {
-            Log.e(TAG, "Sceneform requires OpenGL ES 3.0 or later");
-            Alert.show(getContext(), "Sceneform requires OpenGL ES 3.0 or later");
-        }
-    }
-
-    @Override
-    public View onCreateView(
-            LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        View view = super.onCreateView(inflater, container, savedInstanceState);
-
-        // Turn off the plane discovery since we're only looking for images
-        getPlaneDiscoveryController().hide();
-        getPlaneDiscoveryController().setInstructionView(null);
-        getArSceneView().getPlaneRenderer().setEnabled(false);
+    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
+        view = inflater.inflate(R.layout.fragment_scan, container, false);
+        arFragmentConfig();
         return view;
     }
 
-    @Override
-    protected Config getSessionConfiguration(Session session) {
-        Config config = super.getSessionConfiguration(session);
-        if (!setupAugmentedImageDatabase(config, session)) {
-            Alert.show(getContext(), "Could not setup augmented image database");
-        }
-        return config;
+    private void arFragmentConfig() {
+//        LayoutInflater factory = LayoutInflater.from(getContext());
+//        View view = factory.inflate(R.layout.fragment_scan, null);
+
+        arFragment = (ArFragment) getActivity().getSupportFragmentManager().findFragmentById(R.id.ux_fragment);
+        fitToScanView = view.findViewById(R.id.image_view_fit_to_scan);
+        arFragment.getArSceneView().getScene().addOnUpdateListener(this::onUpdateFrame);
     }
 
-    private boolean setupAugmentedImageDatabase(Config config, Session session) {
-        AugmentedImageDatabase augmentedImageDatabase;
+    private void onUpdateFrame(FrameTime frameTime) {
+        Frame frame = arFragment.getArSceneView().getArFrame();
 
-        AssetManager assetManager = getContext() != null ? getContext().getAssets() : null;
-        if (assetManager == null) {
-            Log.e(TAG, "Context is null, cannot intitialize image database.");
-            return false;
+        // If there is no frame or ARCore is not tracking yet, just return.
+        if (frame == null || frame.getCamera().getTrackingState() != TrackingState.TRACKING) {
+            return;
         }
 
-        // There are two ways to configure an AugmentedImageDatabase:
-        // 1. Add Bitmap to DB directly
-        // 2. Load a pre-built AugmentedImageDatabase
-        // Option 2) has
-        // * shorter setup time
-        // * doesn't require images to be packaged in apk.
-        if (USE_SINGLE_IMAGE) {
-            Bitmap augmentedImageBitmap = loadAugmentedImageBitmap(assetManager);
-            if (augmentedImageBitmap == null) {
-                return false;
+        Collection<AugmentedImage> updatedAugmentedImages =
+                frame.getUpdatedTrackables(AugmentedImage.class);
+        for (AugmentedImage augmentedImage : updatedAugmentedImages) {
+            switch (augmentedImage.getTrackingState()) {
+                case PAUSED:
+                    // When an image is in PAUSED state, but the camera is not PAUSED, it has been detected,
+                    // but not yet tracked.
+                    String text = "Detected Image " + augmentedImage.getIndex();
+                    Alert.show(getContext(), text);
+                    break;
+
+                case TRACKING:
+                    // Have to switch to UI Thread to update View.
+                    fitToScanView.setVisibility(View.GONE);
+
+                    // Create a new anchor for newly found images.
+                    if (!augmentedImageMap.containsKey(augmentedImage)) {
+                        AugmentedImageNode node = new AugmentedImageNode(getContext());
+                        node.setImage(augmentedImage);
+                        augmentedImageMap.put(augmentedImage, node);
+                        arFragment.getArSceneView().getScene().addChild(node);
+                    }
+                    break;
+
+                case STOPPED:
+                    augmentedImageMap.remove(augmentedImage);
+                    break;
             }
-
-            augmentedImageDatabase = new AugmentedImageDatabase(session);
-            augmentedImageDatabase.addImage(DEFAULT_IMAGE_NAME, augmentedImageBitmap);
-            // If the physical size of the image is known, you can instead use:
-            //     augmentedImageDatabase.addImage("image_name", augmentedImageBitmap, widthInMeters);
-            // This will improve the initial detection speed. ARCore will still actively estimate the
-            // physical size of the image as it is viewed from multiple viewpoints.
-        } else {
-            // This is an alternative way to initialize an AugmentedImageDatabase instance,
-            // load a pre-existing augmented image database.
-            try (InputStream is = getContext().getAssets().open(SAMPLE_IMAGE_DATABASE)) {
-                augmentedImageDatabase = AugmentedImageDatabase.deserialize(session, is);
-            } catch (IOException e) {
-                Log.e(TAG, "IO exception loading augmented image database.", e);
-                return false;
-            }
         }
-
-        config.setAugmentedImageDatabase(augmentedImageDatabase);
-        return true;
-    }
-
-    private Bitmap loadAugmentedImageBitmap(AssetManager assetManager) {
-        try (InputStream is = assetManager.open(DEFAULT_IMAGE_NAME)) {
-            return BitmapFactory.decodeStream(is);
-        } catch (IOException e) {
-            Log.e(TAG, "IO exception loading augmented image bitmap.", e);
-        }
-        return null;
     }
 }
