+++ app/src/main/res/layout/fragment_scan.xml
@@ -5,13 +5,8 @@
     android:layout_height="match_parent"
     android:id="@+id/fragment_scan">
 
-    <fragment android:name="com.google.ar.sceneform.ux.ArFragment"
+    <fragment android:name="com.lt3d.fragment.AugmentedImageFragment"
         android:id="@+id/ux_fragment"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent" />
-
-    <com.google.ar.sceneform.ArSceneView
-        android:id="@+id/ar_scene_view"
         android:layout_width="match_parent"
         android:layout_height="match_parent" />
